+++ src/ts/main.ts
@@ -3,6 +3,8 @@ export { version } from './version';
 export * from './UIManager';
 export * from './UIConfig';
 export type { UIComponentConfigMap, UIComponentConfigOverrides } from './UIComponentConfigOverrides';
+export { UIComponentLayoutOverride } from './UIComponentLayoutOverrides';
+export type { UIComponentLayoutOverrideMap, UIComponentLayoutOverrides } from './UIComponentLayoutOverrides';
 // Factories
 export { UIFactory } from './UIFactory';
 // Utils
@@ -109,6 +111,7 @@ export { MetadataLabel, MetadataLabelContent, MetadataLabelConfig } from './comp
 export { AirPlayToggleButton } from './components/buttons/AirPlayToggleButton';
 export { VolumeSlider, VolumeSliderConfig } from './components/seekbar/VolumeSlider';
 export { PictureInPictureToggleButton } from './components/buttons/PictureInPictureToggleButton';
+export { CaptionToggleButton } from './components/buttons/CaptionToggleButton';
 export { Spacer } from './components/Spacer';
 export { BackgroundColorSelectBox } from './components/settings/subtitlesettings/BackgroundColorSelectBox';
 export { BackgroundOpacitySelectBox } from './components/settings/subtitlesettings/BackgroundOpacitySelectBox';